extract and add comment form to person
