+++ location/forms.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from django.forms import ModelForm, HiddenInput, IntegerField, CharField, Form
 
 from core.helper import form_utils
-from core.models import CofkUnionResource, CofkUnionComment
+from core.models import CofkUnionResource
 from location.models import CofkUnionLocation
 from uploader.models import CofkUnionImage
 
@@ -88,31 +88,6 @@ class LocationResourceForm(ModelForm):
             'change_timestamp',
             'change_user',
         )
-
-
-class LocationCommentForm(ModelForm):
-    comment_id = IntegerField(required=False, widget=HiddenInput())
-
-    creation_timestamp = forms.DateTimeField(required=False, widget=HiddenInput())
-    creation_user = forms.CharField(required=False, widget=HiddenInput())
-    change_timestamp = forms.DateTimeField(required=False, widget=HiddenInput())
-    change_user = forms.CharField(required=False, widget=HiddenInput())
-
-    record_tracker_label = form_utils.record_tracker_label_fn_factory('Note')
-
-    class Meta:
-        model = CofkUnionComment
-        fields = (
-            'comment_id',
-            'comment',
-            'creation_timestamp',
-            'creation_user',
-            'change_timestamp',
-            'change_user',
-        )
-        labels = {
-            'comment': 'Note',
-        }
 
 
 class LocUploadImageForm(Form):
