+++ person/templates/person/full_form.html
@@ -78,6 +78,8 @@
 
     </fieldset>
 
+    {% include "core/component/comment_publication.html" with comment_formset=comment_formset %}
+
     <script>
 
         $(function () {
