+++ person/views.py
@@ -1,12 +1,12 @@
 import logging
-from typing import List, Tuple, Callable, Iterable, Type, Optional, Any, NoReturn
+from typing import Callable, Iterable, Type, Optional, Any, NoReturn
 
 from django.db import models
 from django.forms import BaseForm
 from django.shortcuts import render, redirect, get_object_or_404
-from django.views import View
 
-from core.helper import renderer_utils, view_utils
+from core.forms import CommentForm
+from core.helper import renderer_utils, view_utils, model_utils
 from core.helper.view_utils import DefaultSearchView, CommonInitFormViewTemplate
 from location.models import CofkUnionLocation
 from person.forms import PersonForm
@@ -173,6 +173,11 @@ class PersonFullFormHandler:
                                                  name='person_other',
                                                  person=self.person)
 
+        self.comment_formset = view_utils.create_formset(CommentForm, post_data=request_data,
+                                                         prefix='comment',
+                                                         initial_list=model_utils.related_manager_to_dict_list(
+                                                             self.person.comments), )
+
     @property
     def all_recref_handlers(self):
         attr_list = (getattr(self, p) for p in dir(self))
@@ -182,6 +187,7 @@ class PersonFullFormHandler:
     def render_form(self, request):
         context = {
             'person_form': self.person_form,
+            'comment_formset' : self.comment_formset,
         }
         for h in self.all_recref_handlers:
             context.update(h.create_context())
@@ -195,7 +201,7 @@ def full_form(request, iperson_id):
     if request.POST:
 
         # define form_formsets
-        form_formsets = [fhandler.person_form, ]
+        form_formsets = [fhandler.person_form, fhandler.comment_formset]
         for h in fhandler.all_recref_handlers:
             form_formsets.extend([h.new_form, h.update_formset, ])
 
@@ -207,6 +213,9 @@ def full_form(request, iperson_id):
             recref_handler.maintain_record(request, fhandler.person_form.instance)
 
         fhandler.person_form.save()
+        view_utils.save_formset(fhandler.comment_formset,
+                                fhandler.person.comments,
+                                model_id_name='comment_id')
 
         fhandler.load_data(iperson_id, request_data=None)
